Improved source ipv6 compatibility and separated host and port fields
The old source format used square brackets '[' and ']' to delimit the name
component, but square brackets are also used to denote ipv6 addresses. To
avoid ambiguity the name component was moved to the end and the separator
was changed to a single tilda '~' character.

Host and port fields were separated to aid recipient types that might want
to access those values independently.